re-enable catch up
