+++ .github/workflows/continuous-integration.yml
@@ -251,6 +251,7 @@ jobs:
           - test_storage
           - test_synchrony_constraint
           - test_wallets
+          - test_catch_up
           - REMAINDER # Do not modify/remove!
     env:
       TESTS: ${{ matrix.tests }}
